backend/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.1

# 환경변수 및 설정
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF 처리
PyMuPDF>=1.22.0
pdfplumber>=0.9.0
reportlab>=4.0.0

# NLP 및 텍스트 처리
nltk==3.8.1

# 이미지 처리
opencv-python==4.8.1.78
Pillow==10.1.0

# OCR
pytesseract==0.3.10

# 데이터 처리
pandas==2.1.4
numpy==1.24.3

# HTTP 클라이언트
requests==2.31.0
httpx==0.25.2

# AI 및 ML
openai==1.3.7
anthropic==0.7.7

# 로깅 및 디버깅
loguru==0.7.2

# 데이터베이스 마이그레이션
alembic==1.13.1

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1

# 추가된 의존성
aiofiles==23.2.1
pdf2image==1.16.3
pymupdf==1.23.8 